Move skin-related interfaces out of HUD namespace
