+++ osu.Game/Screens/Play/HUD/DefaultScoreCounter.cs
@@ -5,6 +5,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
+using osu.Game.Skinning;
 
 namespace osu.Game.Screens.Play.HUD
 {
