+++ osu.Game/Skinning/ISkinnableTarget.cs
@@ -4,7 +4,7 @@
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 
-namespace osu.Game.Screens.Play.HUD
+namespace osu.Game.Skinning
 {
     /// <summary>
     /// Denotes a container which can house <see cref="ISkinnableComponent"/>s.
